Extract dialog of "TextSetting"
